feat!: Simplify error handling in the child process
Don't try and serialize and send child process errors over the pipe and
panic if that fails.

Instead, maintain a small collection of child process exit statuses that
are interpretable by the parent.

This results in no underlying std::io::Errors being made available from
the child but I think this may be a worthwhile tradeoff for simpler code
and less syscalls/oppertunity for error and ambiguity