test: Remove waitpid_interrupted_by_real_signal() because it doesn't prove the bad case
Or at least it I can't get it to trigger a EINTR from a waitpid(). I
think it has something to do with my libc version or the default
behavior on Debian :shrugs:...

That's even if I revert the "fix" commit (0dd3bf96)...

In any case, _some_ UNIXes should return an EINTR when waitpid is
interrupted by a signal, so I think this branch does bring meaningful
changes, even if I can't reproduce the case on my setup.